+++ src/app/Services/Configuration.php
@@ -44,8 +44,8 @@ class Configuration
 	{
 		return [
 			'domain'         => $this->getEnv('GN_DOMAIN'),
+			'domainSites'    => $this->getEnv('GN_DOMAIN') . '/sites',
 			'debug'          => $this->getEnvBoolean('GN_DEBUG'),
-			'sslAvailable'   => $this->getEnvBoolean('GN_SSL_AVAILABLE'),
 			'rootDir'        => __DIR__ . '/../..',
 			'system_wp'      => $this->getEnvBoolean('GN_SYSTEM_WPCLI'),
 			'custom_wp_path' => $this->getEnv('GN_WPCLI', 'wp'),
@@ -89,6 +89,8 @@ class Configuration
 
 		return [
 			'rootpath'         => "{$root}",
+			'sites'            => "{$root}/sites",
+			'cache'            => "{$root}/cache",
 			'assets'           => "{$root}/assets",
 			'wordpressInstall' => "{$root}/assets/wordpress",
 			'siteExports'      => "{$root}/assets/exports"